introduce lesson 4
