+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/lessons/lesson-2",
     "crates/lessons/lesson-3",
     "crates/lessons/lesson-4",
+    "crates/lessons/lesson-4.5",
 ]
 resolver = "2"
