+++ crates/lessons/lesson-4.5/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "lesson-45"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+macroquad.workspace = true
